+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/memo/controller/MemoController.kt
@@ -1,9 +1,6 @@
-package com.wafflestudio.toyproject.memo_with_tags.memo.controller
+package com.wafflestudio.toyproject.memoWithTags.memo.controller
 
-import com.wafflestudio.toyproject.memo_with_tags.memo.service.MemoService
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestParam
+import com.wafflestudio.toyproject.memoWithTags.memo.service.MemoService
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
