+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/memo/persistence/MemoEntity.kt
@@ -1,0 +1,5 @@
+package com.wafflestudio.toyproject.memoWithTags.memo.persistence
+
+class MemoEntity {
+
+}
