+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/memo/service/MemoService.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.memo_with_tags.memo.service
+package com.wafflestudio.toyproject.memoWithTags.memo.service
 
 import org.springframework.stereotype.Service
