+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/service/TagService.kt
@@ -1,0 +1,4 @@
+package com.wafflestudio.toyproject.memoWithTags.tag.service
+
+class TagService {
+}
