+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/persistence/UserEntity.kt
@@ -1,0 +1,4 @@
+package com.wafflestudio.toyproject.memoWithTags.user.persistence
+
+class UserEntity {
+}
