+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/service/UserService.kt
@@ -1,0 +1,4 @@
+package com.wafflestudio.toyproject.memoWithTags.user.service
+
+class UserService {
+}
