--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/memo/persistence/MemoEntity.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.memo.persistence
-
-import jakarta.persistence.Entity
-
-class MemoEntity {
-
-}
