--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/tag/service/TagService.kt
@@ -1,4 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.tag.service
-
-class TagService {
-}
