--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/user/persistence/UserEntity.kt
@@ -1,4 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.user.persistence
-
-class UserEntity {
-}
